day13: part2
I was stuck on this for a long while. The tests worked, but the actual input didn't seem to exit. I assumed it was a optimisation issue - and starting reading about chinese ramainder.

Then I has a thought that the loop condition might be off, and was stunned when the code exited with the correct answer almost immediately(!)